Kernel: Fix GDT code
The code had various problems. First of all there was an error when
calculating the offset caused by shifting the bits to the left instead
of to the right. Same problem applied to the limit in the GDT Descriptor.
Moreover i was passing a pointer to an array, not taking into account
that an array is already a pointer. I also explicitly disabled the interrupts
via `cli` in the assembly code, and merged the load_gdt and reload_segment_registers
labels into `gdt_flush`, since for whatever reason this seems to work.