:bug: Kernel: Fix a bug in the GDT that was causing the CPU to triple fault or to propagate the error silently
First of all while we were at it we described the best possible way the segment descriptors
via bitfields and enum options. Now it is way clearer to read the code and understand which
flags is being assigned to each segment descriptor. As regards the bug per se it was a mix of two things:
 - not reinterpreting the pointer to the array of descriptors as a dts::u32
 - having the `size` and the `address` swapped in the struct describing the GDT Table
   which was causing one being interpreted as the other