:bug: Kernel: Move GDT to static memory
Having the GDT in a local scope and being destroyed after the function execution
was causing garbage data to be loaded instead. I thought the lgdt instruction
was parsing itself the GDT data structure, but instead I now believe it holds
a pointer to the location where you have stored your GDT in memory.
Took a while to figure it out :(