:rocket: Kernel: Got interrupts up and running
This patch adds an Interrupt Descriptor Table, which is loaded with
CPU exceptions, hardware and software interrupts. Those interrupts are then
handled via their respective assembly routine which calls to a common handler
in C++ which eventually will dispatch the interrupt to the correct hander, once
we have implemented them. We also have setup IRQ (Interrupt Request) and remapped them
within the PIC (Programmable Interrupt Controller) not to conflict with the CPU exceptions.