:rocket: Kernel: Move to paging when allocating memory
This patch includes a basic implementation of paging. There is an API
that would allow us to create new pages and map them to physical memory.
Currently there is an identity mapping for the first 4MB of memory to
preserve GDT, IDT, video memory and the Kernel Code. The latter is remapped
to address space 3GB+ (0xC0000000) effectively creating a Higher Half Kernel.