:bug: Kernel: Fix syscall implementation
Previous to this patch syscall weren't returning the correct value
stored in eax as it was being overwritten by the popa instruction.
Now this has changed and software interrupt 0x80 has its own dedicated
interrupt handler that does not pop eax at the end of the syscall to
preserve the return value.