:rocket: Add new initializer-list constructor for `Vector`
This required a change in the Makfile due to the constraints defined by
the standard. Indeed it is not possible to create a class that will
behave exactly like a `std::initializer_list` since it's not implemented
via language features but rather through compiler magic. The only way we
had to get around this problem was to allow stdlib includes by removing
the `-nostdinc` flag from the build process