:bug: dts: Fix `String::format()` not working on larger strings.
We we're calling a move assignment operator in this function where we
expected the element to be copied. Also, not sure why, passing the
format string by value was causing a page fault. Instead providing the
parameter via a reference (be it an lvalue or rvalue) is working. (still
have to figure this out)

Probably there is still some bug in the current implementation of the
function and as you may see from the source code there are lots of
TODO(s) and FIXME(s).