:recycle: dts: Remove inline assembly from panic macro
The idea behind this is having the possibility to use assert(s) in a
constexpr context. Prior to this patch due to inline assembly it was not
possible. (Even though for whatever reason on my machine it compiled and
ran successfully, whereas on the CI it kept failing). The panic
functionality should remain the same and still halt the machine even
though not directly halting the CPU with the x86_64 instruction.